pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ceramicraft-notification-mservice"
version = "0.1.0"
description = "Push notification microservice for CeramiCraft e-commerce system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dttb>=0.3.2",
    "grpcio>=1.64.1",
    "protobuf>=5.27.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "typer>=0.12.3",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "firebase-admin>=6.5.0",
    "cryptography>=42.0.8",
    "pydantic-settings>=2.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.2",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.6.0",
    "grpcio-tools>=1.64.1",
    "ty>=0.0.22",
]

[project.scripts]
serve = "ceramicraft_notification_mservice.serve:app"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ty.src]
exclude = ["src/ceramicraft_notification_mservice/pb/**"]
